feat(avatar): implement getInitials and getFullName utility functions with tests
