+++ packages/nimbus/src/components/avatar/avatar.tsx
@@ -5,34 +5,7 @@ import { useLocalizedStringFormatter } from "@/hooks";
 import { type AvatarProps } from "./avatar.types";
 import { AvatarRoot } from "./avatar.slots";
 import { avatarMessagesStrings } from "./avatar.messages";
-
-/**
- * Extract a one- or two-character initials string from the supplied names.
- *
- * Defensive against missing input: handles `undefined`, empty strings, and
- * whitespace-only strings by trimming and dropping empty parts. Codepoint-
- * safe: uses `Array.from` so astral-plane characters (e.g. emoji surrogate
- * pairs) are not split mid-surrogate. Returns an empty string when neither
- * input yields a usable character — callers use that signal to render the
- * `Person` icon fallback instead of text.
- */
-export function getInitials(firstName?: string, lastName?: string) {
-  const first = Array.from((firstName ?? "").trim())[0]?.toUpperCase() ?? "";
-  const last = Array.from((lastName ?? "").trim())[0]?.toUpperCase() ?? "";
-  return `${first}${last}`;
-}
-
-/**
- * Compose a display-ready full name from the trimmed non-empty parts.
- * Returns "" when both inputs are missing/empty/whitespace-only so the
- * caller can fall back to a generic localized label.
- */
-export function getFullName(firstName?: string, lastName?: string) {
-  return [firstName, lastName]
-    .map((s) => s?.trim() ?? "")
-    .filter(Boolean)
-    .join(" ");
-}
+import { getInitials, getFullName } from "./utils";
 
 /**
  * Avatar
