+++ packages/nimbus/src/components/avatar/utils/get-initials.spec.ts
@@ -1,14 +1,5 @@
-/**
- * Unit tests for Avatar's pure helpers.
- *
- * Component behavior (rendered DOM, image fallback, icon fallback) is
- * covered by avatar.stories.tsx play functions. These JSDOM-fast tests
- * exercise the initials extraction edge-case matrix that would otherwise
- * require many browser-test permutations.
- */
-
 import { describe, it, expect } from "vitest";
-import { getInitials, getFullName } from "./avatar";
+import { getInitials } from "./get-initials";
 
 describe("getInitials", () => {
   it("returns uppercase first character of each name", () => {
@@ -70,36 +61,5 @@ describe("getInitials", () => {
     // contract.
     expect(typeof getInitials(undefined, undefined)).toBe("string");
     expect(typeof getInitials("", "")).toBe("string");
-  });
-});
-
-describe("getFullName", () => {
-  it("joins both trimmed names with a single space", () => {
-    expect(getFullName("John", "Doe")).toBe("John Doe");
-    expect(getFullName(" John ", " Doe ")).toBe("John Doe");
-  });
-
-  it("returns just firstName when lastName is missing", () => {
-    expect(getFullName("John", undefined)).toBe("John");
-    expect(getFullName("John", "")).toBe("John");
-    expect(getFullName("John", "  ")).toBe("John");
-  });
-
-  it("returns just lastName when firstName is missing", () => {
-    expect(getFullName(undefined, "Doe")).toBe("Doe");
-    expect(getFullName("", "Doe")).toBe("Doe");
-    expect(getFullName(" ", "Doe")).toBe("Doe");
-  });
-
-  it("returns empty string when both are missing", () => {
-    expect(getFullName(undefined, undefined)).toBe("");
-    expect(getFullName("", "")).toBe("");
-    expect(getFullName(" ", "\t")).toBe("");
-  });
-
-  it("does not produce leading, trailing, or doubled spaces", () => {
-    expect(getFullName("John", " ")).toBe("John");
-    expect(getFullName(" ", "Doe")).toBe("Doe");
-    expect(getFullName("  John  ", "  Doe  ")).toBe("John Doe");
   });
 });
