+++ packages/nimbus/src/components/avatar/utils/index.ts
@@ -1,0 +1,2 @@
+export { getInitials } from "./get-initials";
+export { getFullName } from "./get-full-name";
